requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
chromadb>=0.5.0
requests>=2.31.0
pypdf>=4.0.0
gradio>=4.44.0
watchdog>=4.0.0
python-docx>=1.1.0
